Add files via upload
same dealio for q, l and d, q and l were pretty much the same thing just switch the +s and -s. in the Job class there is a function like currentStatus is this just is the job completed or not? because i think based on what needs to be printed the controller class should gather the current status of each of the job